services/realtime-service/requirements.txt
fastapi==0.135.1
starlette==0.49.3
uvicorn==0.34.0
PyJWT==2.12.0
pymongo==4.11.1
prometheus-fastapi-instrumentator==7.0.0
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-exporter-otlp==1.30.0
opentelemetry-instrumentation-fastapi==0.51b0
websockets==13.1
